requirements.txt
Flask==3.1.0
flask_cors==5.0.1
langchain==0.3.0
langchain_community==0.3.0
langchain_core==0.3.1
langchain_google_genai==2.0.0
gunicorn==23.0.0
pydantic==2.9.2
python-dotenv==1.1.0
PyYAML==6.0.1
prophet==1.3.0
